requirements.txt
lyzr-automata==0.1.2
python-dotenv==1.0.1